Break auth component into smaller bits
